+++ pyemma/msm/estimators/maximum_likelihood_msm.py
@@ -187,8 +187,9 @@ class _MSMEstimator(_Estimator, _MSM):
                                     'Consider using sparse=True.'.format(nstates=dtrajstats.nstates))
 
         # count lagged
+        show_progress = getattr(self, 'show_progress', False)
         dtrajstats.count_lagged(self.lag, count_mode=self.count_mode,
-                                mincount_connectivity=self.mincount_connectivity)
+                                mincount_connectivity=self.mincount_connectivity, show_progress=show_progress)
 
         # for other statistics
         return dtrajstats